// frontend/src/types/generated/PathTypeGraph.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PathTypeEdge } from "./PathTypeEdge";
import type { PathTypeNode } from "./PathTypeNode";

/**
 * The OCEL type graph for visualization.
 */
export type PathTypeGraph = { nodes: Array<PathTypeNode>, edges: Array<PathTypeEdge>, };